requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
openai==1.51.2
aiosqlite==0.20.0
duckduckgo-search==6.3.7
jinja2==3.1.4
